Added method prototypes to API
Prototyped objects are created for the convenience of the 'brain' developers. The can become familiar with these functions before they are fully functional and better layout their plan of development for the intelligent systems portion of the project. All functions that are either prototypes or not yet stable have been linked to a dummy method that simply prints that status of the feature.