Added Headlights, removed unneeded ruby file
The Headlights files added are Headlights.cpp, Headlights.h, and
Headlights.js.  These files are completed to the standard available,
and are subject to change if there are needed changes to make the
arduino work.  These will compile, but whether or not these will
work with the arduino is not certain and the javascript commands
are not currently included in commandDispatcher.