Adds headlights interface and moves headlights drivers to drivers folder
The files Headlights.h and Headlights.cpp have been moved to their
appropriate locations and the file IHeadlights.h has been added as
an interface for the headlights.  The command has also been added
to CommandDispatcher to toggle the headlights, and in the Headlights
drivers the Serial class has been implemented for writing to the
arduino.  The files should now be ready to merge.