Changes headlights initialization to be in sync with arduino
This change initializes the headlights' state as already turned
on in order to reflect how the state of the headlights when
the headlights class is initialized.  This is a fix to a bug that
required you to call switchLights twice in order to toggle the lights
for the first time.