Gives basic functionality to ImuSensor and redefines interfaces
ImuSensor is now complete to the point to where it returns accelerometer,
compass, gyroscope and general temperature sensor data.  At this point
it compiles, but tests have not yet been made to ensure that it works.
This currently doesn't give sleep and awake functionality, which is
something that will be added on in the future.